#28: Fix cast issue from 64 to 32 bits on RV32 platforms Use `size_t` for payload length variables, and adapt CCM payload length encoding as `ccmq`

+++ include/api/hardware/v0.5/blockcipher/aes/hca_aes.h
@@ -113,7 +113,7 @@ CRYPTO_FUNCTION int32_t hca_aes_auth_init(
     const metal_scl_t *const scl, aes_auth_ctx_t *const ctx,
     scl_aes_mode_t aes_mode, scl_process_t aes_process,
     scl_endianness_t data_endianness, uint32_t auth_option,
-    const uint8_t *const aad, size_t aad_byte_len, uint64_t payload_len);
+    const uint8_t *const aad, size_t aad_byte_len, size_t payload_len);
 
 /**
  * @brief perform AES cipher with authentication operation
@@ -132,7 +132,7 @@ CRYPTO_FUNCTION int32_t hca_aes_auth_init(
 CRYPTO_FUNCTION int32_t hca_aes_auth_core(const metal_scl_t *const scl,
                                           aes_auth_ctx_t *const ctx,
                                           const uint8_t *const payload,
-                                          uint64_t payload_len,
+                                          size_t payload_len,
                                           uint8_t *const data_out,
                                           size_t *const out_len);
